Don't bother reserving bits
This is a bikeshed colour, but I think a good one.  In the 5 bits that
we have, we can pack two epoch bits (we only need one in practice, but
let's leave that aside for the moment), and a bit each to signal
presence of connection ID and length.  That leaves one spare bit.  I'm
suggesting that we spend this on signaling the difference between an
8-bit and 16-bit sequence number.  16 bits is far more comfortable than
14, and 8 is probably enough for many use cases (saving a byte).

That is, unless you were saving this for a spin bit and VEC.