Remove SENDING->FINISHED transition in state machine
The previous version erroneously specified that implementations
should transition from SENDING to FINISHED after sending the last
flight in a handshake. Instead, they should transition to WAITING
and only move to FINISHED upon receipt of ACKs for the entire last
flight.

This commit fixes the state machine diagram and text.