Fix typo in description of purpose of ACK messages
The initial version

```
The ACK message is used by an endpoint to indicate
handshake-containing the TLS records it has received
from the other side.
```

had a spurious 'the', and it was likely meant to read

```
The ACK message is used by an endpoint to indicate
handshake-containing TLS records it has received
from the other side.
```

instead.

Further, replace 'handshake-containing TLS records' by
'handshake records'.