Consolidate explanations of which records to ACK
There are three sections on ACKs in the current draft:

1. Introduction to ACK messages
   This mainly describes the semantics of a single ACK message,
   and in particular mentions that the only records that
   may be referenced in an ACK message are handshake records
   consisting entirely of handshake messages/fragments which have
   been processed.

2. Sending ACK messages
   This is mainly about which events should be treated as
   signs of disruption in the handshake and hence trigger
   sending ACKs for those handshake messages processed so far.

3. Receiving ACK messages

The current version mixes purposes of 1. and 2. slightly, by
repeating in 2. the statement that ACKs must not be generated
for messages that have been dropped as out-of-order. This belongs
to 1., where it is in fact already stated more generally.

This commit removes the corresponding sentence from 2. and
adds an explicit mentioning of the special case of dropped
out-of-order messages to 1.