Define anti-forgery limits for DTLS
This turned out to be a little tricky.  I copied the text about
AEAD_AES_128_CCM from the QUIC PR, which is easy, but AEAD_AES_128_CCM_8
is a while different story.  As that is defined in TLS 1.3, it seems
necessary to address it here.  QUIC simply doesn't allow that, but I
assume that DTLS will.

On the basis that DTLS will allow AEAD_AES_128_CCM_8, we need some
language about it.  I've tried to write something, but I'm not very
happy with it.  I'd greatly prefer not to allow this AEAD in DTLS.  It's
basically fine in TLS, but here the margins are just too narrow.

Closes #145.